+++ src/main/kotlin/sea/hack/club/graphql/types/event/EventTagRatingType.kt
@@ -3,6 +3,6 @@ package sea.hack.club.graphql.types.event
 import sea.hack.club.graphql.types.TagRatingType
 
 data class EventTagRatingType(
-        val eventId: Long,
+        val eventId: Int,
         val ratings: List<TagRatingType>
 )
