+++ src/main/kotlin/sea/hack/club/graphql/types/tag/TagType.kt
@@ -2,9 +2,9 @@ package sea.hack.club.graphql.types.tag
 
 import sea.hack.club.entity.Skill
 
-data class TagType(val id: Long)
+data class TagType(val id: Int)
 
 fun Skill.toGraphType(): TagType {
     val id = checkNotNull(this.id)
-    return TagType(id)
+    return TagType(id.toInt())
 }
